fix(ps_core): 修复图层恢复与文本属性修改逻辑
- 将恢复图层失败的日志级别从 info 提升为 error
- 增加对 textItem 属性的检查，确保字体大小和颜色在需要时恢复到初始状态
- 添加 restore_text_item_to_initial 方法用于恢复文本属性
- 调整日志输出内容，增强调试信息可读性
- 完善测试用例，增加对字体修改场景的覆盖
- 更新依赖版本，包括 comtypes、lxml、psutil 和 pywin32
- 调整项目配置，统一管理测试相关设置
