+++ pyproject.toml
@@ -5,10 +5,6 @@ description = "This is a  Python  for Photoshop."
 readme = "README.md"
 requires-python = ">=3.13"
 dependencies = [
-    "photoshop-python-api>=0.24.1",
-]
-
-[dependency-groups]
-dev = [
-    "xlwings>=0.33.15",
+    "photoshop-python-api",
+    "xlwings"
 ]
